limit test patches to specific component, move atches into conftest
