+++ tests/components/onkyo/conftest.py
@@ -1,23 +1,14 @@
 """Configure tests for the Onkyo integration."""
 
-from collections.abc import Generator
-from unittest.mock import AsyncMock, patch
+from unittest.mock import patch
 
 import pytest
 
 from homeassistant.components.onkyo.const import DOMAIN
 
+from . import create_connection
+
 from tests.common import MockConfigEntry
-
-
-@pytest.fixture
-def mock_setup_entry() -> Generator[AsyncMock]:
-    """Override async_setup_entry."""
-    with patch(
-        "homeassistant.components.onkyo.async_setup_entry",
-        return_value=True,
-    ) as mock_setup_entry:
-        yield mock_setup_entry
 
 
 @pytest.fixture(name="config_entry")
@@ -28,3 +19,59 @@ def mock_config_entry() -> MockConfigEntry:
         title="Onkyo",
         data={},
     )
+
+
+@pytest.fixture(autouse=True)
+def patch_timeouts():
+    """Patch timeouts to avoid tests waiting."""
+    with (
+        patch(
+            "homeassistant.components.onkyo.receiver.DEVICE_INTERVIEW_TIMEOUT", new=0
+        ),
+        patch(
+            "homeassistant.components.onkyo.receiver.DEVICE_DISCOVERY_TIMEOUT", new=0
+        ),
+    ):
+        yield
+
+
+@pytest.fixture
+async def default_mock_discovery():
+    """Mock discovery with a single device."""
+
+    async def mock_discover(host=None, discovery_callback=None, timeout=0):
+        await discovery_callback(create_connection(1))
+
+    with patch(
+        "homeassistant.components.onkyo.receiver.pyeiscp.Connection.discover",
+        new=mock_discover,
+    ):
+        yield
+
+
+@pytest.fixture
+async def stub_mock_discovery():
+    """Mock discovery with no devices."""
+
+    async def mock_discover(host=None, discovery_callback=None, timeout=0):
+        pass
+
+    with patch(
+        "homeassistant.components.onkyo.receiver.pyeiscp.Connection.discover",
+        new=mock_discover,
+    ):
+        yield
+
+
+@pytest.fixture
+async def empty_mock_discovery():
+    """Mock discovery with an empty connection."""
+
+    async def mock_discover(host=None, discovery_callback=None, timeout=0):
+        await discovery_callback(None)
+
+    with patch(
+        "homeassistant.components.onkyo.receiver.pyeiscp.Connection.discover",
+        new=mock_discover,
+    ):
+        yield
